+++ backend/smody/src/main/java/com/woowacourse/smody/image/strategy/ImgBBImageStrategy.java
@@ -1,8 +1,5 @@
 package com.woowacourse.smody.image.strategy;
 
-import com.woowacourse.smody.exception.BusinessException;
-import com.woowacourse.smody.exception.ExceptionData;
-import com.woowacourse.smody.image.dto.ImageUrlResponse;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -12,10 +9,19 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.woowacourse.smody.exception.BusinessException;
+import com.woowacourse.smody.exception.ExceptionData;
+import com.woowacourse.smody.image.dto.ImageUrlResponse;
+
+import lombok.RequiredArgsConstructor;
+
 @Component
+@RequiredArgsConstructor
 public class ImgBBImageStrategy implements ImageStrategy {
 
     private static final String IMGBB_KEY = "085936a4cafeb33b7d1a80073e9a1f91";
+
+    private final RestTemplate restTemplate;
 
     @Override
     public String extractUrl(MultipartFile rawImage) {
@@ -30,7 +36,7 @@ public class ImgBBImageStrategy implements ImageStrategy {
         headers.setContentType(MediaType.MULTIPART_FORM_DATA);
 
         HttpEntity<MultiValueMap<String, Object>> httpEntity = new HttpEntity<>(body, headers);
-        ImageUrlResponse response = new RestTemplate()
+        ImageUrlResponse response = restTemplate
                 .postForEntity(requestUri, httpEntity, ImageUrlResponse.class)
                 .getBody();
         if (response == null) {
