+++ backend/smody/src/main/java/com/woowacourse/smody/image/strategy/SmodyImageStrategy.java
@@ -1,7 +1,5 @@
 package com.woowacourse.smody.image.strategy;
 
-import com.woowacourse.smody.exception.BusinessException;
-import com.woowacourse.smody.exception.ExceptionData;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Primary;
 import org.springframework.http.HttpEntity;
@@ -14,17 +12,24 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.woowacourse.smody.exception.BusinessException;
+import com.woowacourse.smody.exception.ExceptionData;
+
+import lombok.RequiredArgsConstructor;
+
 @Component
+@RequiredArgsConstructor
 @Primary
 public class SmodyImageStrategy implements ImageStrategy {
 
     @Value("${secret.key.upload}")
     private String secretKeyUpload;
+    private final RestTemplate restTemplate;
 
     @Override
     public String extractUrl(MultipartFile rawImage) {
         validateEmptyImage(rawImage);
-        ResponseEntity<String> imageServerResponse = new RestTemplate()
+        ResponseEntity<String> imageServerResponse = restTemplate
                 .postForEntity("https://images.smody.co.kr/images/upload", generateImageRequest(rawImage),
                         String.class);
         if (imageServerResponse.getStatusCode().is4xxClientError()) {
